Add the ability to encrypt a file
This is a larger change than I had intended to create, mainly because
encryption isn't my area of expertise, nor is it an area that I'm really
all that familiar with. So this change is what it is. :-(

The change refactors App to store the keyphrase, nonce, and gcm so that
they're the same throughout the application, and available to all of
App's methods.

Given that, encryptFileData is now a lot smaller as most of the
functionality has been moved to the NewApp function. However, the new,
complimentary decryptFileData method is as small as it can be because it
only needs to decrypt the supplied byte array and return it, or an
error, if the text could not be decrypted.

Thanks to these changes, the EncryptFile method has been completed, and
covering tests have been added to cover most of the available
functionality. I don't know that I've covered all of the paths through
the code, yet.

A big thank you has to go to
https://earthly.dev/blog/cryptography-encryption-in-go/ and
https://stackoverflow.com/a/11531951/222011 where I found the core of
the implementation (which was then refactored) and how to actually test
the implementation.