Add the ability to decrypt encrypted text
This change adds a handler and route for decrypting encrypted text. I'm
no guru or authority on encryption, but have come to understand just
enough about how it works to get it to work (plus a bit). The encrypted
text needs to include the nonce as a prefix for it to be able to be
decrypted. Consequently, the encryption functions have been, slightly,
refactored to include the nonce as a prefix in the encrypted text.