Encrypt test sent in the request, not the contents of a file
This change refactors the code so that it encrypts text sent in the body
of the request, not the contents of an uploaded file. Originally, my
thoughts on how the application would work weren't that clear, so the
approach that I took was quite muddled, assuming that a file's contents
would be encrypted. Over the course of developing the code to date, I
learned more and now have a clearer understanding of what the app should
do. So, this refactor steps it further in that direction by removing
the need to retrieve the plain text to encrypt from an uploaded file,
rather from a parameter in the body of the request.