Better organise the code
The intention of this change is to better organise the code so that it's
easier to work with over the course of time (allowing for this code
being used for more than a short period of time).

So, to that end, this change is the first, of likely a few, that
organise the code into a series of packages (I believe that's the
correct term). I'm somewhat new to creating my own packages, so please
excuse any incorrect terminology and rawness as I get into use them.